Create separate query popup for redemption and use existing one for opening+preview
